Add try finally

+++ tests/marc_exporter/test_marc_export.py
@@ -118,35 +118,44 @@ class TestMarcExport(unittest.TestCase):
             }
         }
 
-        # export marc
-        issue, articles = HelperTestMethods.create_custom_issue(issue, [article1, article2])
-        marc_exporter.main(issue['_id'], config.default_elastic_index)
+        try:
+            # export marc
+            issue, articles = HelperTestMethods.create_custom_issue(issue, [article1, article2])
+            marc_exporter.main(issue['_id'], config.default_elastic_index)
 
-        #  check if files exists
-        marc_issue = Path(path_issue + "/issue_marc21.txt")
-        self.assertEqual(True, marc_issue.is_file(), "No file " + path_issue + "/issue_marc21.txt")
-        marc1 = Path(path_issue + "/articles/1/1_marc21.txt")
-        self.assertEqual(True, marc1.is_file(), "No file " + path_issue + "/articles/1/1_marc21.txt")
-        marc2 = Path(path_issue + "/articles/2/2_marc21.txt")
-        self.assertEqual(True, marc2.is_file(), "No file " + path_issue + "/articles/2/2_marc21.txt")
+            #  check if files exists
+            marc_issue = Path(path_issue + "/issue_marc21.txt")
+            self.assertEqual(True, marc_issue.is_file(), "No file " + path_issue + "/issue_marc21.txt")
+            marc1 = Path(path_issue + "/articles/1/1_marc21.txt")
+            self.assertEqual(True, marc1.is_file(), "No file " + path_issue + "/articles/1/1_marc21.txt")
+            marc2 = Path(path_issue + "/articles/2/2_marc21.txt")
+            self.assertEqual(True, marc2.is_file(), "No file " + path_issue + "/articles/2/2_marc21.txt")
 
-        es = Elasticsearch()
-        issue = es.get(index='deep_search_test_python', doc_type='issue', id=issue['_id'])
-        articles = es.search(index='deep_search_test_python', doc_type="article",
-                             body={'query': {'bool': {'must': {
-                                 'nested': {'path': 'issue', 'query': {'match': {'issue.id': issue['_id']}}}}}},
-                                 'size': 1000})['hits']['hits']
+            es = Elasticsearch()
+            issue = es.get(index='deep_search_test_python', doc_type='issue', id=issue['_id'])
+            articles = es.search(index='deep_search_test_python', doc_type="article",
+                                 body={'query': {'bool': {'must': {
+                                     'nested': {'path': 'issue', 'query': {'match': {'issue.id': issue['_id']}}}}}},
+                                     'size': 1000})['hits']['hits']
 
-        self.assertEqual(True, issue['_source']['journal_marc21_path'] == path)
-        self.assertEqual(True, issue['_source']['issue_marc21_path'] == path_issue + "/issue_marc21.txt")
-        for article in articles:
-            self.assertEqual(True, article['_source']['article_marc21_path'] == path_issue + "/articles/1/1_marc21.txt"
-                             or article['_source']['article_marc21_path'] == path_issue + "/articles/2/2_marc21.txt")
+            self.assertEqual(True, issue['_source']['journal_marc21_path'] == path)
+            self.assertEqual(True, issue['_source']['issue_marc21_path'] == path_issue + "/issue_marc21.txt")
+            for article in articles:
+                self.assertEqual(True, article['_source']['article_marc21_path'] == path_issue + "/articles/1/1_marc21.txt"
+                                 or article['_source']['article_marc21_path'] == path_issue + "/articles/2/2_marc21.txt")
 
-        #  delete files
-        os.remove(path_issue + "/issue_marc21.txt")
-        shutil.rmtree(path_issue + "/articles")
-        os.remove(issue['_source']['journal_marc21_path'])
+        finally:
+            #  delete files
+            marc_issue = Path(path_issue + "/issue_marc21.txt")
+            if marc_issue.is_file():
+                os.remove(path_issue + "/issue_marc21.txt")
+            marc1 = Path(path_issue + "/articles/1/1_marc21.txt")
+            marc2 = Path(path_issue + "/articles/2/2_marc21.txt")
+            if marc1.is_file() or marc2.is_file():
+                shutil.rmtree(path_issue + "/articles")
+            marc_journal = Path(issue['_source']['journal_marc21_path'])
+            if marc_journal.is_file():
+                os.remove(issue['_source']['journal_marc21_path'])
 
 
 if __name__ == '__main__':
